refactor: unify Context::add_jit_cost on u64
Consolidates `add_jit_cost` (u32) and `add_jit_cost_u64` into a single
`add_jit_cost(amount: u64)` with saturating addition. Since this API was
introduced in PR 854 and has no pre-existing external users, the u32/u64
split carried no real benefit — it was just unnecessary ceremony around
"is this contribution bounded by input structure or by tx structure".
The u64 version's saturating-add is safer than the u32 version's
wrapping-add (well-formed inputs can't hit it, but pathological cases
could), so unify on that semantics.

Ripple changes:

* `Context::add_per_item_jit_cost` internally computes cost as u32
  (base + chunks * per_chunk — the product is bounded by small cost
  constants × collection sizes) and widens to u64 at the final
  `add_jit_cost(u64::from(cost))` call. Keeps the u32 argument types
  for call-site ergonomics.
* Per-opcode cost constants that were `u32` are now `u64`:
  `EVAL_SIGMA_PROP_CONSTANT` in `eval.rs`, and the ten
  `EQ_*_COST` / `COLL_MATCH_TYPE_COST` constants in `data_value_comparer.rs`.
  No literal call sites change — integer literals infer as u64 from the
  param type.
* `tx_context.rs::validate()`: `add_jit_cost_u64(init_cost_jit)` becomes
  plain `add_jit_cost(init_cost_jit)`; the crypto-cost path drops the
  `u32::try_from` dance entirely and passes `estimate_crypto_cost`'s
  u64 return value straight through. The overflow-to-CostLimitExceeded
  fallback is now covered by the unified `add_jit_cost`'s own limit
  check — if the sum exceeds the limit, it returns `CostLimitExceeded`
  directly. Removes two unused imports (`CostLimitExceeded`, `EvalError`).

Behavior-preserving: cost-parity smoke harness still 78/78 vs Scala,
Phase 1 cumulative-limit regression passes unchanged, full workspace
green, C bindings build clean.

+++ ergotree-interpreter/src/eval.rs
@@ -131,7 +131,7 @@ pub struct ReductionResult {
 /// JIT cost for a script that trivially reduces to a SigmaProp constant (e.g.
 /// bare P2PK). Scala's `EvalSigmaPropConstant` charges 50 JitCost; pre-fix we
 /// only paid the generic `Expr::Const` cost of 5 JitCost.
-const EVAL_SIGMA_PROP_CONSTANT: u32 = 50;
+const EVAL_SIGMA_PROP_CONSTANT: u64 = 50;
 
 /// Short-circuit for trees whose proposition is a plain SigmaProp constant.
 /// Returns `Some(sigma_bool)` for both forms:
